lisa: tests: scheduler: Move test_slack in eas_behaviour
